+++ src/main/java/ro/msg/learning/shop/controllers/StockExportController.java
@@ -5,6 +5,8 @@ import org.springframework.web.bind.annotation.*;
 import ro.msg.learning.shop.dto.StockDTO;
 import ro.msg.learning.shop.services.StockExportService;
 
+import javax.servlet.http.HttpServletResponse;
+import java.net.http.HttpHeaders;
 import java.util.List;
 
 
@@ -15,10 +17,11 @@ public class StockExportController {
 
     private final StockExportService stockExportService;
 
-//    value = "/goods/{locationId}", produces = "text/plain; charset=utf-8"
+//    value = "/stocks/{locationId}", produces = "text/plain; charset=utf-8"
 
-    @GetMapping(value = "/goods/{locationId}", produces = "text/plain; charset=utf-8")
-    public List<StockDTO> exporting(@PathVariable int locationId) {
+    @GetMapping(value = "/stocks/{locationId}", produces = "text/csv")
+    public List<StockDTO> exporting(@PathVariable int locationId, HttpServletResponse response) {
+//        response.setHeader(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=stocks.csv");
         return stockExportService.exporting(locationId);
     }
 }
